Removed old classes. SwipeDismissCallback usable by other recyclers.
