+++ Joey/UI/Fragments/LogTimeEntriesListFragment.cs
@@ -2,13 +2,10 @@
 using System.Collections.Generic;
 using Android.Animation;
 using Android.Content;
-using Android.Graphics;
-using Android.Graphics.Drawables;
 using Android.OS;
 using Android.Support.V4.App;
 using Android.Support.V7.Widget;
 using Android.Support.V7.Widget.Helper;
-using Android.Util;
 using Android.Views;
 using Android.Widget;
 using Toggl.Joey.Data;
@@ -24,7 +21,7 @@ using XPlatUtils;
 
 namespace Toggl.Joey.UI.Fragments
 {
-    public class LogTimeEntriesListFragment : Fragment, SwipeDismissTouchListener.IDismissCallbacks, ItemTouchListener.IItemTouchListener
+    public class LogTimeEntriesListFragment : Fragment, SwipeDismissCallback.IDismissListener, ItemTouchListener.IItemTouchListener
     {
         private static readonly int UndbarVisibleTime = 6000;
         private static readonly int UndbarScrollThreshold = 100;
@@ -118,7 +115,7 @@ namespace Toggl.Joey.UI.Fragments
             itemTouchListener = new ItemTouchListener (recyclerView, this);
             recyclerView.AddOnItemTouchListener (itemTouchListener);
 
-            var touchCallback = new RecyclerSwipeCallback (ItemTouchHelper.Up | ItemTouchHelper.Down, ItemTouchHelper.Left);
+            var touchCallback = new SwipeDismissCallback (ItemTouchHelper.Up | ItemTouchHelper.Down, ItemTouchHelper.Left, this);
             var touchHelper = new ItemTouchHelper (touchCallback);
             touchHelper.AttachToRecyclerView (recyclerView);
 
@@ -162,18 +159,18 @@ namespace Toggl.Joey.UI.Fragments
             }
         }
 
-        #region IDismissCallbacks implementation
+        #region IDismissListener implementation
 
-        public bool CanDismiss (RecyclerView view, int position)
+        public bool CanDismiss (RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder)
         {
-            var adapter = view.GetAdapter ();
-            return adapter.GetItemViewType (position) == LogTimeEntriesAdapter.ViewTypeContent;
+            var adapter = recyclerView.GetAdapter ();
+            return adapter.GetItemViewType (viewHolder.LayoutPosition) == LogTimeEntriesAdapter.ViewTypeContent;
         }
 
-        public void OnDismiss (RecyclerView view, int position)
+        public void OnDismiss (RecyclerView.ViewHolder viewHolder, int position)
         {
             var undoAdapter = recyclerView.GetAdapter () as IUndoCapabilities;
-            undoAdapter.RemoveItemWithUndo (position);
+            undoAdapter.RemoveItemWithUndo (viewHolder);
             ShowUndoBar ();
         }
 
@@ -199,7 +196,8 @@ namespace Toggl.Joey.UI.Fragments
 
         public bool CanClick (RecyclerView view, int position)
         {
-            return CanDismiss (view, position);
+            var adapter = recyclerView.GetAdapter ();
+            return adapter.GetItemViewType (position) == LogTimeEntriesAdapter.ViewTypeContent;
         }
 
         #endregion
@@ -297,116 +295,6 @@ namespace Toggl.Joey.UI.Fragments
                 if (owner.UndoBarVisible) {
                     owner.RemoveItemAndHideUndoBar ();
                 }
-            }
-        }
-
-        private class RecyclerSwipeCallback : ItemTouchHelper.SimpleCallback
-        {
-            private const float minThreshold = 20;
-
-            private int leftBorderWidth;
-            private Drawable backgroundShape;
-            private Context ctx;
-            private string deleteText;
-            private Paint labelPaint;
-            private Rect rect = new Rect();
-
-            private Context Ctx
-            {
-                get {
-                    if (ctx == null) {
-                        ctx = ServiceContainer.Resolve<Context> ();
-                    }
-                    return ctx;
-                }
-            }
-
-            private string DeleteText
-            {
-                get {
-                    if (string.IsNullOrEmpty (deleteText)) {
-                        deleteText = ctx.Resources.GetString (Resource.String.SwipeDeleteQuestion);
-                    }
-                    return deleteText;
-                }
-            }
-
-            private Drawable BackgroundShape
-            {
-                get {
-                    if (backgroundShape == null) {
-                        backgroundShape = Ctx.Resources.GetDrawable (Resource.Drawable.swipe_background_shape);
-                    }
-                    return backgroundShape;
-                }
-            }
-
-            private Paint LabelPaint
-            {
-                get {
-                    if (labelPaint == null) {
-                        var labelFontSize = TypedValue.ApplyDimension (ComplexUnitType.Sp, 14, ctx.Resources.DisplayMetrics);
-                        labelPaint = new Paint {
-                            Color = Color.White,
-                            TextSize = labelFontSize,
-                            AntiAlias = true,
-                        };
-                        labelPaint.GetTextBounds (DeleteText, 0, DeleteText.Length, rect);
-                        leftBorderWidth = (int)TypedValue.ApplyDimension (ComplexUnitType.Dip, 32, ctx.Resources.DisplayMetrics);
-                    }
-                    return labelPaint;
-                }
-            }
-
-            public RecyclerSwipeCallback (IntPtr a, Android.Runtime.JniHandleOwnership b) : base (a, b)
-            {
-            }
-
-            public RecyclerSwipeCallback (int p0, int p1) : base (p0, p1)
-            {
-            }
-
-            public override bool OnMove (RecyclerView p0, RecyclerView.ViewHolder p1, RecyclerView.ViewHolder p2)
-            {
-                return false;
-            }
-
-            public override int GetSwipeDirs (RecyclerView p0, RecyclerView.ViewHolder p1)
-            {
-                var adapter = ((LogTimeEntriesAdapter)p0.GetAdapter ());
-                if (adapter.GetItemViewType (p1.AdapterPosition) == LogTimeEntriesAdapter.ViewTypeContent) {
-                    return ItemTouchHelper.Right;
-                }
-                return 0;
-            }
-
-            public override void OnSwiped (RecyclerView.ViewHolder p0, int p1)
-            {
-            }
-
-            public override float GetSwipeThreshold (RecyclerView.ViewHolder p0)
-            {
-                return minThreshold;
-            }
-
-            public override void OnChildDraw (Canvas c, RecyclerView recyclerView, RecyclerView.ViewHolder viewHolder, float dX, float dY, int actionState, bool isCurrentlyActive)
-            {
-                var itemHeight = viewHolder.ItemView.Height;
-                BackgroundShape.SetBounds (0, (int)viewHolder.ItemView.GetY (), c.Width, (int)viewHolder.ItemView.GetY () + itemHeight);
-                BackgroundShape.Draw (c);
-                c.DrawText (DeleteText, leftBorderWidth, viewHolder.ItemView.GetY () + itemHeight /2.0f + rect.Height ()/2.0f, LabelPaint);
-
-                base.OnChildDraw (c, recyclerView, viewHolder, dX, dY, actionState, isCurrentlyActive);
-            }
-
-            protected override void Dispose (bool disposing)
-            {
-                if (disposing && backgroundShape != null) {
-                    backgroundShape.Dispose ();
-                    labelPaint.Dispose ();
-                }
-
-                base.Dispose (disposing);
             }
         }
     }
